% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_nearby_features}
\alias{annotate_nearby_features}
\title{Annotates a GRanges object with nearby features in another GRanges object}
\usage{
annotate_nearby_features(
  gr,
  feat_gr,
  name_field,
  distance_cutoff = 2500,
  ignore.strand = TRUE
)
}
\arguments{
\item{gr}{GRanges object to be annotated}

\item{feat_gr}{GRanges features to be annotated.}

\item{name_field}{Name of the column that contains the annotation name in feat_gr}

\item{distance_cutoff}{Maximum distance around the promoter of the feature}

\item{ignore.strand}{Ignore strand when annotating. Default FALSE}
}
\value{
A GRanges object with a nearby_features field in it containing names
of overlapping features in that distance range (separated by commas)
}
\description{
Distance is measured as distance from end and start depending whether the
target locus is downstream or upstream.
}
\examples{

features_gr <- GenomicRanges::GRanges(
  seqnames = c("chr1", "chr1"),
  IRanges::IRanges(c(10,22), c(20,30)),
  strand = c("-", "-"),
  name = c("Feat_A", "Feat_B")
)

gr <- GenomicRanges::GRanges(
  seqnames = c("chr1"),
  IRanges::IRanges(15, 25),
  strand = "+"
)

annotate_nearby_features(gr, features_gr, "name", ignore.strand = TRUE)
}
