% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_nearest_features}
\alias{annotate_nearest_features}
\title{Annotate the closest feature}
\usage{
annotate_nearest_features(gr, feat_gr, name_field, ignore.strand = TRUE)
}
\arguments{
\item{gr}{Query GRanges object to annotate}

\item{feat_gr}{Features to annotate it with}

\item{name_field}{Name of the column in feat_gr to get the name from}

\item{ignore.strand}{Whether to report loci from any strand or only matching.
By default it is set to true.}
}
\value{
gr with extra column "nearby_features" annotated
}
\description{
This works similar as annotate nearby features but is
}
\examples{

features_gr <- GenomicRanges::GRanges(
  seqnames = c("chr1", "chr1"),
  IRanges::IRanges(c(10,22), c(20,30)),
  strand = c("-", "-"),
  name = c("Feat_A", "Feat_B")
)

gr <- GenomicRanges::GRanges(
  seqnames = c("chr1"),
  IRanges::IRanges(15, 25),
  strand = "+"
)

annotate_nearest_features(gr, features_gr, "name")
}
