% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{calculate_venn_intersections}
\alias{calculate_venn_intersections}
\title{Calculates all intersections in a list of GRanges object for a venn diagram
of an arbitrary number of elements.}
\usage{
calculate_venn_intersections(grlist, names = NULL, ignore.strand = TRUE)
}
\arguments{
\item{grlist}{List of GRanges objects (maximum 4)}

\item{names}{identifiers for the GRanges objects.}

\item{ignore.strand}{If FALSE overlaps will be counted only with matching strand}
}
\value{
A named array for all intersections named as names.
}
\description{
Note that this can be computationally demanding (subsets grow exponentially)
and probably pairwise intersections would be of more interest for larger
number of sets, so there is a hard limit on 5 sets.
}
\details{
NOTE: Calculates intersection values as number of loci in the intersection.
}
\examples{
gr_1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(10, 20), strand = "-")
gr_2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 25), strand = "+")
gr_3 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(16, 18), strand = "+")
intersection <- calculate_venn_intersections(
  list(gr_1, gr_2, gr_3),
  names = c("a", "b", "c"),
  ignore.strand = TRUE
)
}
