% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{combinations_enrichment}
\alias{combinations_enrichment}
\title{Calculate enrichment combinations of different elements within a list}
\usage{
combinations_enrichment(grlist, genome_size, ignore.strand = TRUE)
}
\arguments{
\item{grlist}{Named list of GRanges objects}

\item{genome_size}{Genome size}

\item{ignore.strand}{If true, strand is not considered for overlaps.}
}
\value{
A data frame in long format with the enrichment values calculated
per pair.
}
\description{
Calculate enrichment combinations of different elements within a list
}
\examples{
gr_a1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(11, 20), strand = "-")
gr_a2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(24, 25), strand = "+")
gr_b1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 24), strand = "-")

enrichment <- combinations_enrichment(
  list("A" = gr_a1, "B" = gr_a2, "X" = gr_b1),
  genome_size = 30,
  ignore.strand = FALSE
)
}
