% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{combinations_score}
\alias{combinations_score}
\title{Apply a given score function to all pairs in a list of GRanges objects}
\usage{
combinations_score(grlist, score_func)
}
\arguments{
\item{grlist}{named List of GRanges objects}

\item{score_func}{Any scoring function (must take a pair of GRanges objects)}
}
\value{
A data frame
}
\description{
Apply a given score function to all pairs in a list of GRanges objects
}
\examples{
gr_a1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(11, 20), strand = "-")
gr_a2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(24, 25), strand = "+")
gr_b1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 24), strand = "-")

# Fix the function parameters
score_func <- purrr::partial(
  fc_enrichment,
  genome_size = genome_size,
  ignore.strand = ignore.strand
)

combinations_score
  list("A" = gr_a1, "B" = gr_a2, "X" = gr_b1),
  score_func
)
}
