% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{.intersect_subset}
\alias{.intersect_subset}
\title{Intersection of a list of GRanges filtering by index.
The use of this is to be able to map by powerset}
\usage{
.intersect_subset(grlist, indices, ignore.strand)
}
\arguments{
\item{grlist}{List of GRanges}

\item{indices}{Array of integers}

\item{ignore.strand}{Whether to ignore strand in the intersection}
}
\value{
Size of the intersection between grlist[indices]
}
\description{
Intersection of a list of GRanges filtering by index.
The use of this is to be able to map by powerset
}
