% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{.make_venn_names}
\alias{.make_venn_names}
\title{Make names for intersections according to index list}
\usage{
.make_venn_names(indices, names)
}
\arguments{
\item{indices}{Indices to be used}

\item{names}{Names list}
}
\value{
names[indices] joined by an "&" character
}
\description{
Make names for intersections according to index list
}
