% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{.powerset}
\alias{.powerset}
\title{Calculates all possible subsets of a list of elements}
\usage{
.powerset(set)
}
\arguments{
\item{set}{List of elements}
}
\value{
A list of vectors
}
\description{
Calculates all possible subsets of a list of elements
}
