% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fit_euler}
\alias{fit_euler}
\title{Find a fit for a set of intersections in the form of an euler object.}
\usage{
fit_euler(grlist, names = NULL, ignore.strand = TRUE, shape = "circle")
}
\arguments{
\item{grlist}{List of GRanges objects}

\item{names}{Names to give the sets}

\item{ignore.strand}{Whether to ignore strand for intersections}

\item{shape}{Which shape to use, either circle or ellipse. Circle is the
default, but ellipse is more likely to be accurate for a difficult fit.}
}
\value{
An euler object
}
\description{
Find a fit for a set of intersections in the form of an euler object.
}
\examples{
gr_1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(10, 20), strand = "-")
gr_2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 25), strand = "+")
gr_3 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(16, 18), strand = "+")
fit_euler(list(gr_1, gr_2, gr_3), names = c("a", "b", "c"))
}
