% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{gr_universe}
\alias{gr_universe}
\title{GRanges universe for a set of conditions and replicates}
\usage{
gr_universe(
  grlist,
  min_consensus,
  resize_pre = NULL,
  resize_across = NULL,
  resize_post = NULL,
  anchor = "center"
)
}
\arguments{
\item{grlist}{A list where each element is a list of GRanges objects}

\item{min_consensus}{Minimum consensus needed to merge peaks from the same condition (replicates)}

\item{resize_pre}{Resize replicate peaks before calling consensus - Default NULL}

\item{resize_across}{Resize peaks before merging across conditions.
Resizing to a small size here will make sure that proximal peaks will not
get merged, but might incur in certain level of overlap if resize_post is
larger.}

\item{resize_post}{Resize final group of peaks to a fixed size. Default NULL}

\item{anchor}{Where to anchor for resizing. Defaults to center.}
}
\description{
GRanges universe for a set of conditions and replicates
}
