% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{grlist_intersect_size}
\alias{grlist_intersect_size}
\title{Calculate the size of the intersection of a list of GRanges objects.}
\usage{
grlist_intersect_size(grlist, ignore.strand = TRUE)
}
\arguments{
\item{grlist}{A list of GRanges}

\item{ignore.strand}{If TRUE, overlaps are counted without taking strand into
account.}
}
\description{
Calculate the size of the intersection of a list of GRanges objects.
}
\examples{
gr_1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(10, 20), strand = "-")
gr_2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 25), strand = "+")
gr_3 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(16, 18), strand = "+")
intersection <- grlist_intersect_size(list(gr_1, gr_2, gr_3), ignore.strand = TRUE)
}
