% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_named_bed_into_list}
\alias{import_named_bed_into_list}
\title{Make a named bed into a list of GRanges objects to be used with pairwise
functions, mostly.}
\usage{
import_named_bed_into_list(bedpath)
}
\arguments{
\item{bedpath}{A path to a bedfile}
}
\value{
A list of GRanges objects
}
\description{
Make a named bed into a list of GRanges objects to be used with pairwise
functions, mostly.
}
