% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{impute_feature}
\alias{impute_feature}
\title{Annotate each locus in a GRanges with the single best overlap in another
feature GRanges. Best overlap is calculated as jaccard index:}
\usage{
impute_feature(
  gr,
  feature_gr,
  name_field,
  minoverlap = 1L,
  ignore.strand = TRUE
)
}
\arguments{
\item{gr}{GRanges to annotate}

\item{feature_gr}{GRanges with features - Needs to have a name field}

\item{name_field}{Name of the column that contains the feature names in feature_gr}

\item{minoverlap}{Minimum overlap to consider.}

\item{ignore.strand}{Whether to ignore strand on annotation.}
}
\value{
GRanges gr annotated with a "feature" metadata column with the name
of the feature that best overlaps with it, NA if none do.
}
\description{
length(intersect) / length(union)
}
\examples{

features_gr <- GenomicRanges::GRanges(
  seqnames = c("chr1", "chr1"),
  IRanges::IRanges(c(10,22), c(20,30)),
  strand = c("-", "-"),
  name = c("Feat_A", "Feat_B")
)

gr <- GenomicRanges::GRanges(
  seqnames = c("chr1"),
  IRanges::IRanges(15, 25),
  strand = "+"
)

impute_feature(gr, features_gr, "name", ignore.strand = TRUE)
}
