% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{jaccard_index}
\alias{jaccard_index}
\title{Calculate jaccard index between two GenomicRanges objects.}
\usage{
jaccard_index(gr1, gr2, ignore.strand = TRUE)
}
\arguments{
\item{gr1}{GRanges object}

\item{gr2}{GRanges object}

\item{ignore.strand}{If FALSE, only matching strand overlaps are counted}
}
\value{
Numeric value representing the jaccard intersection
}
\description{
Jaccard index = length(intersection) / length(union)
}
\examples{
gr_1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(10, 20), strand = "-")
gr_2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 25), strand = "+")
jaccard_index(gr_1, gr_2, ignore.strand = TRUE)
}
