% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{loci_overlap}
\alias{loci_overlap}
\title{Calculate the number of loci that overlap between two GRanges objects
(ignoring length).}
\usage{
loci_overlap(gr1, gr2, ignore.strand = TRUE, minoverlap = 1L)
}
\arguments{
\item{gr1}{A GRanges object}

\item{gr2}{A GRanges object}

\item{ignore.strand}{If FALSE, not matching strands will not be counted.}

\item{minoverlap}{Minimum overlap in bp to consider this an overlap}
}
\value{
An integer total number of loci in gr1 that overlap with any locus in gr2
}
\description{
Note that this is *NOT a commutative operation, since one locus in gr1 could
overlap with > 1 locus in gr2, and viceversa. It is what is used generally
on venn diagrams, which is why those do not usually add up completely.
However, if there is not highly fragmentation difference, results should not
be extremely different.
}
\examples{
gr1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(11, 20), strand = "-")
gr2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(24, 25), strand = "+")

loci_overlap(gr1, gr2)
}
