% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise_enrichment}
\alias{pairwise_enrichment}
\title{Calculate pairwise enrichment matrix between two lists of GRanges objects}
\usage{
pairwise_enrichment(grlist_1, grlist_2, genome_size, ignore.strand = TRUE)
}
\arguments{
\item{grlist_1}{A named list of GRanges}

\item{grlist_2}{A named list of GRanges}

\item{genome_size}{Genome size}

\item{ignore.strand}{Ignore strand to calculate overlaps.}
}
\value{
A data frame in long format with the enrichment values calculated
per pair.
}
\description{
Calculate pairwise enrichment matrix between two lists of GRanges objects
}
\examples{
gr_a1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(11, 20), strand = "-")
gr_a2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(24, 25), strand = "+")

gr_b1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 24), strand = "-")
gr_b2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(16, 25), strand = "+")

gsize <- 30

enrichment <- pairwise_enrichment(
  list("A" = gr_a1, "B" = gr_a2),
  list("X" = gr_b1, "Y" = gr_b2),
  genome_size = gsize,
  ignore.strand = FALSE
)
}
