% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_combinations_enrichment}
\alias{plot_combinations_enrichment}
\title{Plot triangular pairwise enrichment heatmap for a single GRanges list}
\usage{
plot_combinations_enrichment(grlist, genome_size, ignore.strand = TRUE)
}
\arguments{
\item{grlist}{Named list of GRanges objects}

\item{genome_size}{Genome size}

\item{ignore.strand}{If true, strand is not considered for overlaps.}
}
\value{
A ggplot object
}
\description{
Plot triangular pairwise enrichment heatmap for a single GRanges list
}
\examples{
gr_a1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(11, 20), strand = "-")
gr_a2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(24, 25), strand = "+")

gr_b1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 24), strand = "-")
gr_b2 <- GenomicRanges::GRanges(seqnames = c("chr3"), IRanges::IRanges(16, 25), strand = "+")

plot_combinations_enrichment(
  list("A" = gr_a1, "B" = gr_a2, "X" = gr_b1, "Y" = gr_b2),
  genome_size = 30,
  ignore.strand = FALSE
)
}
