% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_euler}
\alias{plot_euler}
\title{Plot a Euler diagram with intersections between GRanges objects}
\usage{
plot_euler(grlist, names = NULL, ignore.strand = TRUE, fills = NULL)
}
\arguments{
\item{grlist}{List of GRanges objects}

\item{names}{Names to give the sets}

\item{ignore.strand}{Whether to ignore strand for intersections}

\item{fills}{Color fills for each set. Eulerr interpolates the intersections}
}
\value{
A euler.diagram object that can be further processed, or plotted via print, ggsave
}
\description{
Further parameters can be passed to the plot function (see plot.euler)
}
\examples{
gr_1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(10, 20), strand = "-")
gr_2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 25), strand = "+")
gr_3 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(16, 18), strand = "+")
plot_euler(list(gr_1, gr_2, gr_3), names = c("a", "b", "c"))
}
