% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pairwise_enrichment}
\alias{plot_pairwise_enrichment}
\title{Plot pairwise enrichment heatmap}
\usage{
plot_pairwise_enrichment(grlist_1, grlist_2, genome_size, ignore.strand = TRUE)
}
\arguments{
\item{grlist_1}{A named list of GRanges}

\item{grlist_2}{A named list of GRanges}

\item{genome_size}{Genome size}

\item{ignore.strand}{Ignore strand to calculate overlaps.}
}
\value{
ggplot object
}
\description{
Plot pairwise enrichment heatmap
}
\examples{
gr_a1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(11, 20), strand = "-")
gr_a2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(24, 25), strand = "+")

gr_b1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 24), strand = "-")
gr_b2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(16, 25), strand = "+")

gsize <- 30

plot_pairwise_enrichment(
  list("A" = gr_a1, "B" = gr_a2),
  list("X" = gr_b1, "Y" = gr_b2),
  genome_size = gsize,
  ignore.strand = FALSE
)
}
