% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{total_bp_overlap}
\alias{total_bp_overlap}
\title{Total overlapping base pairs between two GRanges objects.}
\usage{
total_bp_overlap(gr1, gr2, ignore.strand = TRUE)
}
\arguments{
\item{gr1}{GRanges object}

\item{gr2}{GRanges object}

\item{ignore.strand}{Whether to ignore strand when calculating overlap.}
}
\value{
Integer representing the total bp overlap between two sets of loci.
}
\description{
Total overlapping base pairs between two GRanges objects.
}
\examples{
gr_1 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(10, 20), strand = "-")
gr_2 <- GenomicRanges::GRanges(seqnames = c("chr1"), IRanges::IRanges(15, 25), strand = "+")
total_bp_overlap(gr_1, gr_2, ignore.strand = TRUE)
total_bp_overlap(gr_1, gr_2, ignore.strand = FALSE)
}
